==> 1_Abstract.ipynb <==
### Einleitung

Um das Wissen, welches wir im Modul Big Data Lab Cluster erlernten, zu vertiefen, konnten wir eine Projektarbeit durchführen. Diese umfasste den kompletten Prozess vom Erstellen eines Clusters bis zur Auswertung der Daten.

Zu Beginn stand die Konfiguration unserer drei virtuellen Maschinen für den Clusterverbund an, gleichzeitig machten wir uns auch bereits auf die Suche, welches Datenset wir auswerten möchten. Dabei stiessen wir unter www.kaggle.com  auf das «Airline on-time Performance Data» Datenset, welches die Flugdaten der Inlandflüge in den USA von 1987 bis 2008 beinhaltet. Wir entschieden uns schnell für dieses Datenset, mit welchem wir die Hauptfrage «Ist im Datenset der Einfluss von 9-11 im Flugverkehr quantifizierbar?» beantworten wollen. 

Das Datenset von www.kaggle.com wurde mittels der API von Kaggle heruntergeladen, um es anschliessend im HDFS von Hadoop verteilt zu speichern. Danach wurde das Schema des Datenset inspiziert und als Parquet gespeichert. Mit SPARK machten wir uns anschliessend an die Analyse des >12GB grossen Datenset. Wie es uns dabei ergangen ist und wie gross die Veränderungen im Flugverkehr nach dem Terroranschlag waren, kann in den weiteren Abschnitten entnommen werden.
