==> 4_Prototyping.ipynb <==
# Prototyping

import findspark
from pyspark.sql import SparkSession
from pyspark.sql.functions import col, isnan, when, count
from pyspark.sql.types import StructType, StructField, IntegerType, StringType, BooleanType

findspark.init()
spark = SparkSession.builder.appName('Prototyping').getOrCreate()

## Carriers File

df_carriers = spark.read.parquet('/airline-data/carriers.parquet')

Gemäss Kaggle sollte das File zwei Spalten (Code und Description) enthalten sowie 1491 Zeilen haben. Alle Zeilen sollten Distinct sein.

df_carriers.printSchema()

df_carriers.count()

df_carriers.select('Code').distinct().count()

df_carriers.select('Description').distinct().count()

Da Count() und distinct().count() dieselben Werte liefern, können wir sicherstellen, dass keine Duplikate vorhanden sind.

df_carriers.select([count(when(isnan(c) | col(c).isNull(), c)).alias(c) for c in df_carriers.columns]).show()

Das Carriers Parquet File ist somit einwandfrei. Das heisst es hat in den zwei Spalten keine null-Werte.

## Airlinedata File

df_airlines = spark.read.parquet('/airline-data/airline.parquet')

Das Airlines File sollte gemäss Kaggle 29 Spalten haben und nach unserer Überprüfung 123'534'969 Zeilen.

df_airlines.printSchema()

Beim Erstellen des Parquet-Formats hatten wir das Problem, dass die Schema-Inferenz nicht auf Parquet übernommen wurde, sondern alle Spalten als "String" erstellt wurden.

len(df_airlines.columns)

df_airlines.count()

Die Anzahl Spalten und Zeilen stimmen: 29 und 123'534'969

df_airlines.select([count(when(isnan(c) | col(c).isNull(), c)).alias(c) for c in df_airlines.columns]).show(truncate=False)

Es hat 139'774 Nullwerte in der Spalte TailNum d.H. es gibt keine Angabe zur Tail Number (das Nummernschild von Flugzeugen). In der Spalte CancelationCode hat es 38'955'823 Nullwerte. Entweder sind die Flüge nicht gecancelled worden, oder es gibt für diesen Fall keinen Code.

df_airlines.filter(col('CancellationCode').isNull()).groupBy('Cancelled').count().show()

Diese Abfrage bestätigt unsere Vermutung, dass bei allen "Cancelled"-Flüge (Cancelled = 1) beim CancellationCode ein null steht. Somit macht das Datenset Sinn, dass wenn ein Flug nicht gecancelled wurde, somit auch NULL beim CancellationCodes steht.

spark.stop()

## Erläuterungen

### Hat es Überraschungen beim ersten Kontakt mit den Daten gegeben?

In fast einem Drittel der Zeilen hat es Nullwerte. Bei genauerem Hinschauen sind die aber nicht schlimm, resp. zu erwarten. siehe z.Bsp. Ausführungen oben bezüglich Cancellation-Code.

### Was haben wir aus dem prototyping gelernt?

Beim Einlesen der Daten aus .csv Files funktioniert die Schemainferenz nicht richtig. Alle Daten werden als Strings erkannt, obwohl das meiste eigentlich Integers sein müssten. Korrekterweise würden wir die Spalten zum richtigen Typ casten, damit die SQL Operationen effizienter ablaufen. Auf diesen Zusatzschritt haben wir verzeichtet.    
Ansonsten sind die Daten in Ordnung und auch die Nullwerte ergeben Sinn.
