==> 5_Dataflow.ipynb <==
# Dataflow

## Wie werden die Daten End-to-End durch das System laufen?

Die Daten werden als Raw Files direkt über die Kaggle-API auf die Master Node geladen, entpackt und dann unverändert ins HDFS geladen. Aus den Raw Files im HDFS werden dann zwei Parquet Files erstellt und die Originale werden vom HDFS wieder gelöscht. Die Parquet Files werden in der Analyse entsprechend durch SPARK ausgelesen.

## Wie werden die Rohdaten gespeichert?

Die Rohdaten werden nur kurz auf dem Master und später im HDFS zwischengespeichert. Sobald die Rohdaten im platzsparenderem Parquet Format gespeichert sind, werden sie gelöscht.

## Sind die Daten partitioniert? Wenn ja, weshalb und wie?

Die Parquet Files sind mit dem Partitionsfaktor 55 gespeichert. Dies entspricht einer generellen Empfehlung, den Partitionsfaktor so zu wählen, dass er 4 x der verfügbaren Cores entspricht. Die Partitionierung wurde beim Erstellen der Parquet Files gemacht.

## In welchem Format sind die prozessierten Daten gespeichert? Und wieso?

Die prozessierten Daten sind als Parquet Files gespeichert. Diese Files sind dank der effizienten Kompression sehr platzsparend und können schnell und effizient gelesen werden.
