==> 8_Quellen.ipynb <==
# Quellen

Für die Arbeit haben wir folgende Plattformen verwendet:

SPARK uns SPARK-SQL Dokumentation: https://spark.apache.org/docs/latest/api/python/reference/index.html

Matplotlib: https://stackoverflow.com/questions/48062499/matplotlib-y-axis-values-are-not-ordered

Round within SPARK: https://stackoverflow.com/questions/47046827/trouble-with-pyspark-round-function

Parquet File: https://spark.apache.org/docs/latest/sql-data-sources-parquet.html

Lesen und Schreiben von Parquet-Files: https://sparkbyexamples.com/spark/spark-read-write-dataframe-parquet-example/

Null und NaN mit PaSpark: https://sparkbyexamples.com/pyspark/pyspark-find-count-of-null-none-nan-values/

Parquet / Schema:
 - https://stackoverflow.com/questions/65132280/write-parquet-from-another-parquet-with-a-new-schema-using-pyspark
 - https://stackoverflow.com/questions/52387021/spark-read-parquet-with-custom-schema
 - https://stackoverflow.com/questions/50625165/change-schema-of-existing-dataframe

und vieles mehr über Google (kein Code kopiert sondern verstanden)
